requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI and LangChain
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector Store
pinecone-client==3.0.0

# Telegram
python-telegram-bot==20.7

# Utilities
aiohttp==3.9.1
requests==2.31.0
python-multipart==0.0.6

# Data processing
pandas==2.1.4
numpy==1.26.2

# Logging and monitoring
colorlog==6.8.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0
pre-commit==3.6.0

# Deployment
gunicorn==21.2.0
uvicorn==0.25.0
fastapi==0.108.0

# Docker support
psutil==5.9.6
